+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Unreleased
+
+- Test samples have been moved into password protected zips to protect them from a local antivirus
+- Added a test signature unique to the mock as an alternative to the eicar signature
+- Ruby and dependency updates
+
 ## V1.3.0
 
 - Ruby and dependency updates